requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.2.0
scikit-learn>=1.4.0
numpy>=1.26.3
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
python-dotenv>=1.0.1
matplotlib>=3.8.2
seaborn>=0.13.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pillow>=10.2.0
groq>=0.11.0
httpx>=0.27.0
